feat(users): refactor AdminUsersController import and update UsersModule to reflect new file structure
