+++ backend/src/users/users.controller.ts
@@ -1,10 +1,5 @@
-import { Controller, Post, Get, Patch, Param, Body, UseGuards, Req } from '@nestjs/common';
+import { Controller, Get, UseGuards, Req } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
-import { AuthGuard } from '@nestjs/passport';
-import { RolesGuard } from '../auth/roles.guard';
-import { Roles } from '../auth/roles.decorator';
 import { JwtAuthGuard } from '../auth/jwt-auth.guard';
 import { IdentityResolutionService } from '../auth/identity-resolution.service';
 import { Request } from 'express';
@@ -63,34 +58,5 @@ export class UsersController {
         updated_at: assignment.updated_at,
       })),
     };
-  }
-}
-
-@Controller('admin/users')
-@UseGuards(AuthGuard('jwt'), RolesGuard)
-@Roles('admin')
-export class AdminUsersController {
-  constructor(private readonly usersService: UsersService) {}
-
-  @Post()
-  async create(@Body() dto: CreateUserDto) {
-    const user = await this.usersService.create(dto);
-    return { message: 'User created successfully', user };
-  }
-
-  @Get()
-  async list() {
-    return this.usersService.findAll();
-  }
-
-  @Get(':id')
-  async get(@Param('id') id: string) {
-    return this.usersService.findOne(id);
-  }
-
-  @Patch(':id')
-  async update(@Param('id') id: string, @Body() dto: UpdateUserDto) {
-    const user = await this.usersService.update(id, dto);
-    return { message: 'User updated successfully', user };
   }
 }
