+++ backend/src/users/users.module.ts
@@ -5,7 +5,8 @@ import { Role } from '../roles/role.entity';
 import { Entity as OrgEntity } from '../entities/entity.entity';
 import { UserRoleAssignment } from '../roles/user-role-assignment.entity';
 import { UsersService } from './users.service';
-import { UsersController, AdminUsersController } from './users.controller';
+import { UsersController } from './users.controller';
+import { AdminUsersController } from './admin-users.controller';
 import { IdentityResolutionService } from '../auth/identity-resolution.service';
 import { IdpIdentity } from '../idp-identities/idp-identity.entity';
 import { RolesModule } from '../roles/roles.module';
